🐛 Fix the URLs generated by docs.help_icon on the progress_effects page. (#207)
Without the `with context` the generated URL is relative to the octoprint server, as opposed to being on https://cp2004.gitbook.io/ws281x-led-status/.

It seems this was a mistake when this change was [originally made](https://github.com/bramp/OctoPrint-WS281x_LED_Status/commit/3648aa58815db2289a761b1cb14ed59486f32139). Additionally it seems overview.jinga2 has a `with context` where it doesn't need to be.